Renamed FBAInputGrid to FBATable
