+++ web/src/features/fbaCalculator/components/WeatherStationCell.tsx
@@ -6,7 +6,7 @@ import React, { useEffect, useState } from 'react'
 import {
   FBAInputGridProps,
   GridMenuOption
-} from 'features/fbaCalculator/components/FBAInputGrid'
+} from 'features/fbaCalculator/components/FBATable'
 import { buildUpdatedOptionRow, updateFBARow } from 'features/fbaCalculator/tableState'
 
 interface WeatherStationCellProps {
