+++ web/src/features/fbaCalculator/components/grassCureCell.test.tsx
@@ -1,5 +1,5 @@
 import { render, screen } from '@testing-library/react'
-import { FBAInputRow } from 'features/fbaCalculator/components/FBAInputGrid'
+import { FBAInputRow } from 'features/fbaCalculator/components/FBATable'
 import GrassCureCell, {
   GrassCureCellProps
 } from 'features/fbaCalculator/components/GrassCureCell'
