+++ web/src/features/fbaCalculator/components/windSpeedCell.test.tsx
@@ -1,5 +1,5 @@
 import { render, screen } from '@testing-library/react'
-import { FBAInputRow } from 'features/fbaCalculator/components/FBAInputGrid'
+import { FBAInputRow } from 'features/fbaCalculator/components/FBATable'
 import WindSpeedCell, {
   WindSpeedCellProps
 } from 'features/fbaCalculator/components/WindSpeedCell'
